+++ src/cli/actions/backup.rs
@@ -42,6 +42,7 @@ impl Runner {
         let act_restore = arg_command == "restore";
         let act_rmhome = arg_command == "rmhome";
         let act_rmbackup = arg_command == "rmbackup";
+        let profile = self.load_profile(1)?;
         let wflag_all = self.args.flags().contains(&Flag::Word("all".into()));
         let lflag_all = self.args.flags().contains(&Flag::Letter('a'));
         let flag_all = wflag_all || lflag_all;
@@ -55,10 +56,11 @@ impl Runner {
 
         // resolve profile into all leafs
         let mut profile_loader = Self::profile_loader()?;
-        let root_profile = profile_loader.load(&self.load_profile(1)?)?;
+        let root_profile = profile_loader.load(&profile)?;
         let profiles = root_profile.resolve(&mut profile_loader)?;
 
         // iterate over all leaf profiles
+        self.output_main_profile(&profile);
         for profile in profiles {
             match profile.ptype() {
                 ProfileType::Composite(_) => unreachable!("Composite profile impossible here"),